perf(circuit): index qubits with u32 rather than usize
Qubit and CBit were aliases for usize, which made Gate 32 bytes: ccx
carries three operands, and the discriminant rounds the rest up. At u32
a Gate is 16, halving the footprint of the one data structure the whole
pipeline is built on. On the largest cobble-rz circuits -- around twelve
million gates -- that is the difference between holding a circuit copy
in 378 MB and in 189 MB, and every pass holds several.

Counts stay usize; only indices change. Per-qubit arrays cast at the
point of indexing, and everything that carries a qubit is Qubit-typed
end to end, so the casts sit at array subscripts rather than being
sprinkled through the passes. The QASM resolvers now return Qubit and
CBit directly, which puts the one conversion from untrusted input in a
single place -- and with it a bound: a qreg or creg wider than u32::MAX
is a parse error instead of a silent truncation.

On ols-ridge-d20 at -O3, min-of-3 interleaved: 20.97s -> 17.20s
sequential and 10.33s -> 7.41s with --parallel, both cumulative with the
preceding CnotMin commit. Peak RSS falls from 3.68 GB to 2.19 GB
sequential, 3.00 GB to 1.96 GB parallel, and across the rest of the
corpus from 3.1-3.7 GB to 1.5-2.2 GB. The allocator effectively leaves
the profile: malloc goes from 8.1% of samples to 0.2%, and what remains
is the SuperOpt window scan at 40.6% self.

Callers passing a usize qubit index now need `as u32`; integer literals
are unaffected.

Output is byte-identical on all 19 corpus files under 25 MB and on the
five large ones, at -O1/-O2/-O3, --decompose-cz and --parallel.

+++ src/optimize.rs
@@ -248,12 +248,12 @@ impl Metrics {
             let (arity, operands) = qubit_operands(gate);
             let layer = operands[..arity]
                 .iter()
-                .map(|&qubit| next_layer[qubit])
+                .map(|&qubit| next_layer[qubit as usize])
                 .max()
                 .unwrap_or(0)
                 + 1;
             for &qubit in &operands[..arity] {
-                next_layer[qubit] = layer;
+                next_layer[qubit as usize] = layer;
             }
             depth = depth.max(layer);
         }
